Smart_Sniffer/requirements.txt
# Smart Sniffer - AV Cabin Air Quality Monitor
# Python dependencies for Arduino + host serial workflow

# Core dependency
pyserial>=3.5          # Serial communication with Arduino BME688 stream

# Optional legacy dependency (Raspberry Pi I2C path)
# smbus2>=0.4.0

# Development/Testing
pytest>=7.0.0
pytest-cov>=4.0.0
